+++ CHANGES.rst
@@ -4,8 +4,11 @@ Changelog
 0.38 (unreleased)
 -----------------
 
-- Nothing changed yet.
-
+- Fixed code to work with `plone.app.async` as in this case, there is no
+ `REQUEST`.  To do this, needed to get the `@@images` view by instantiating the
+  `ImageDataModifiedImageScaling` class, this could be a problem if it is
+  overrided by a subpackage.
+  [gbastien]
 
 0.37 (2019-06-14)
 -----------------